+++ src/app.ts
@@ -26,9 +26,7 @@ import {
   CGN_API_CLIENT,
   CGN_OPERATOR_SEARCH_API_CLIENT,
   ENV,
-  EUCOVIDCERT_API_CLIENT,
   FF_CGN_ENABLED,
-  FF_EUCOVIDCERT_ENABLED,
   FF_IO_FIMS_ENABLED,
   FF_IO_SIGN_ENABLED,
   FF_IO_WALLET_ENABLED,
@@ -51,6 +49,8 @@ import {
   NOTIFICATIONS_QUEUE_NAME,
   NOTIFICATIONS_STORAGE_CONNECTION_STRING,
   PAGOPA_CLIENT,
+  PAGOPA_ECOMMERCE_CLIENT,
+  PAGOPA_ECOMMERCE_UAT_CLIENT,
   PN_ADDRESS_BOOK_CLIENT_SELECTOR,
   PNAddressBookConfig,
   PUSH_NOTIFICATIONS_QUEUE_NAME,
@@ -64,7 +64,6 @@ import {
   registerCgnAPIRoutes,
   registerCgnOperatorSearchAPIRoutes
 } from "./routes/cgnRoutes";
-import { registerEUCovidCertAPIRoutes } from "./routes/euCovidCertRoutes";
 import { registerFirstLollipopConsumer } from "./routes/firstLollipopConsumerRoutes";
 import { registerIoFimsAPIRoutes } from "./routes/ioFimsRoutes";
 import { registerIoSignAPIRoutes } from "./routes/ioSignRoutes";
@@ -77,7 +76,6 @@ import { registerTrialSystemAPIRoutes } from "./routes/trialSystemRoutes";
 import AuthenticationLockService from "./services/authenticationLockService";
 import CgnOperatorSearchService from "./services/cgnOperatorSearchService";
 import CgnService from "./services/cgnService";
-import EUCovidCertService from "./services/eucovidcertService";
 import IoFimsService from "./services/fimsService";
 import FunctionsAppService from "./services/functionAppService";
 import IoSignService from "./services/ioSignService";
@@ -86,6 +84,7 @@ import LollipopService from "./services/lollipopService";
 import NewMessagesService from "./services/newMessagesService";
 import NotificationService from "./services/notificationService";
 import { getNotificationServiceFactory } from "./services/notificationServiceFactory";
+import PagoPAEcommerceService from "./services/pagoPAEcommerceService";
 import PagoPAProxyService from "./services/pagoPAProxyService";
 import { PNService } from "./services/pnService";
 import ProfileService from "./services/profileService";
@@ -119,7 +118,6 @@ export interface IAppFactoryParameters {
   readonly IoSignAPIBasePath: string;
   readonly IoFimsAPIBasePath: string;
   readonly IoWalletAPIBasePath: string;
-  readonly EUCovidCertBasePath: string;
   readonly ServicesAppBackendBasePath: string;
   readonly TrialSystemBasePath: string;
 }
@@ -135,7 +133,6 @@ export async function newApp({
   IoFimsAPIBasePath,
   IoWalletAPIBasePath,
   CGNOperatorSearchAPIBasePath,
-  EUCovidCertBasePath,
   ServicesAppBackendBasePath,
   TrialSystemBasePath
 }: IAppFactoryParameters): Promise<Express> {
@@ -288,11 +285,6 @@ export async function newApp({
           CGN_OPERATOR_SEARCH_API_CLIENT
         );
 
-        // Create the EUCovidCert service
-        const EUCOVIDCERT_SERVICE = new EUCovidCertService(
-          EUCOVIDCERT_API_CLIENT
-        );
-
         // Create the user data processing service
         const USER_DATA_PROCESSING_SERVICE = new UserDataProcessingService(
           API_CLIENT
@@ -386,6 +378,12 @@ export async function newApp({
         );
 
         const PAGOPA_PROXY_SERVICE = new PagoPAProxyService(PAGOPA_CLIENT);
+
+        const PAGOPA_ECOMMERCE_SERVICE = new PagoPAEcommerceService(
+          PAGOPA_ECOMMERCE_CLIENT,
+          PAGOPA_ECOMMERCE_UAT_CLIENT
+        );
+
         registerAPIRoutes(
           app,
           APIBasePath,
@@ -397,6 +395,7 @@ export async function newApp({
           notificationServiceFactory,
           SESSION_STORAGE,
           PAGOPA_PROXY_SERVICE,
+          PAGOPA_ECOMMERCE_SERVICE,
           USER_DATA_PROCESSING_SERVICE,
           authMiddlewares.bearerSession,
           LOLLIPOP_API_CLIENT
@@ -446,15 +445,6 @@ export async function newApp({
             IoFimsAPIBasePath,
             IO_FIMS_SERVICE,
             PROFILE_SERVICE,
-            authMiddlewares.bearerSession
-          );
-        }
-
-        if (FF_EUCOVIDCERT_ENABLED) {
-          registerEUCovidCertAPIRoutes(
-            app,
-            EUCovidCertBasePath,
-            EUCOVIDCERT_SERVICE,
             authMiddlewares.bearerSession
           );
         }