+++ src/services/__tests__/ioWalletService.test.ts
@@ -437,11 +437,15 @@ describe("IoWalletService#createWalletAttestationV2", () => {
   it("should make the correct api call", async () => {
     const service = new IoWalletService(api, trialSystemApi);
 
-    await service.createWalletAttestationV2("assertion" as NonEmptyString);
+    await service.createWalletAttestationV2(
+      "assertion" as NonEmptyString,
+      aFiscalCode
+    );
 
     expect(mockCreateWalletAttestationV2).toHaveBeenCalledWith({
       body: {
-        assertion: "assertion"
+        assertion: "assertion",
+        fiscal_code: aFiscalCode
       }
     });
   });
@@ -450,7 +454,8 @@ describe("IoWalletService#createWalletAttestationV2", () => {
     const service = new IoWalletService(api, trialSystemApi);
 
     const res = await service.createWalletAttestationV2(
-      "assertion" as NonEmptyString
+      "assertion" as NonEmptyString,
+      aFiscalCode
     );
 
     expect(res).toMatchObject({
@@ -466,7 +471,8 @@ describe("IoWalletService#createWalletAttestationV2", () => {
     const service = new IoWalletService(api, trialSystemApi);
 
     const res = await service.createWalletAttestationV2(
-      "assertion" as NonEmptyString
+      "assertion" as NonEmptyString,
+      aFiscalCode
     );
 
     expect(res).toMatchObject({
@@ -483,7 +489,8 @@ describe("IoWalletService#createWalletAttestationV2", () => {
     const service = new IoWalletService(api, trialSystemApi);
 
     const res = await service.createWalletAttestationV2(
-      "assertion" as NonEmptyString
+      "assertion" as NonEmptyString,
+      aFiscalCode
     );
 
     expect(res).toMatchObject({
@@ -500,7 +507,8 @@ describe("IoWalletService#createWalletAttestationV2", () => {
     const service = new IoWalletService(api, trialSystemApi);
 
     const res = await service.createWalletAttestationV2(
-      "assertion" as NonEmptyString
+      "assertion" as NonEmptyString,
+      aFiscalCode
     );
 
     expect(res).toMatchObject({
@@ -515,7 +523,8 @@ describe("IoWalletService#createWalletAttestationV2", () => {
     const service = new IoWalletService(api, trialSystemApi);
 
     const res = await service.createWalletAttestationV2(
-      "assertion" as NonEmptyString
+      "assertion" as NonEmptyString,
+      aFiscalCode
     );
 
     expect(res).toMatchObject({